// FrontEnd/src/interfaces/IMovieDetails.ts
export default interface IMovieDetails {
    "countries": string[];
    "description": string;
    "directors": string[];
    "genres": string[];
    "imdb_id": string;
    "imdb_rating": string;
    "language": string[];
    "popularity": string;
    "rated": string;
    "release_date": string;
    "runtime": number;
    "stars": string[];
    "title": string;
    "vote_count": string;
    "year": string;
    "youtube_trailer_key": string;
    "fanart": string;
    "poster": string;
}